MAINT: refactor project structure
