+++ src/piqtree2/__init__.py
@@ -1,21 +1,21 @@
 """piqtree2 - access the power of IQ-TREE within Python."""
 
-from ._iq_wrappers import (
+from piqtree2.iqtree import (
     TreeGenMode,
     build_tree,
     fit_tree,
     random_trees,
     robinson_foulds,
 )
-from ._options import available_models
+from piqtree2.models import available_models
 
 __version__ = "0.1.1.dev0"
 
 __all__ = [
+    "available_models",
     "build_tree",
     "fit_tree",
     "random_trees",
     "robinson_foulds",
     "TreeGenMode",
-    "available_models",
 ]
