+++ src/piqtree2/_libiqtree/README.txt
@@ -1,0 +1,1 @@
+Place the IQ-TREE static library, `libiqtree2.a`, in this directory.
