+++ src/piqtree2/iqtree/_decorators.py
@@ -7,9 +7,8 @@ import tempfile
 from functools import wraps
 from typing import Callable, Optional, TypeVar
 
-from typing_extensions import ParamSpec
-
 from piqtree2.exceptions import IqTreeError
+from typing_extensions import ParamSpec
 
 Param = ParamSpec("Param")
 RetType = TypeVar("RetType")
