+++ src/piqtree2/models/__init__.py
@@ -1,0 +1,3 @@
+from ._options import available_models
+
+__all__ = ["available_models"]
